[Core] Change Sequence to ulong
Fix #896, #875

+++ Lagrange.Core/Internal/Context/Logic/Implementation/OperationLogic.cs
@@ -39,7 +39,7 @@ internal class OperationLogic : LogicBase
 
     public async Task<MessageResult> SendMessage(MessageChain chain)
     {
-        uint clientSeq = chain.ClientSequence;
+        ulong clientSeq = chain.ClientSequence;
         ulong messageId = chain.MessageId;
 
         var sendMessageEvent = SendMessageEvent.Create(chain);
@@ -601,7 +601,7 @@ internal class OperationLogic : LogicBase
         return results.Count != 0 ? ((MultiMsgUploadEvent)results[0]).ResId : null;
     }
 
-    public async Task<bool> MarkAsRead(uint groupUin, string? targetUid, uint startSequence, uint time)
+    public async Task<bool> MarkAsRead(uint groupUin, string? targetUid, ulong startSequence, uint time)
     {
         var markAsReadEvent = MarkReadedEvent.Create(groupUin, targetUid, startSequence, time);
         var results = await Collection.Business.SendEvent(markAsReadEvent);
@@ -614,14 +614,14 @@ internal class OperationLogic : LogicBase
         return results.Count != 0 && results[0].ResultCode == 0;
     }
 
-    public async Task<bool> SetEssenceMessage(uint groupUin, uint sequence, uint random)
+    public async Task<bool> SetEssenceMessage(uint groupUin, ulong sequence, uint random)
     {
         var setEssenceMessageEvent = SetEssenceMessageEvent.Create(groupUin, sequence, random);
         var results = await Collection.Business.SendEvent(setEssenceMessageEvent);
         return results.Count != 0 && ((SetEssenceMessageEvent)results[0]).ResultCode == 0;
     }
 
-    public async Task<bool> RemoveEssenceMessage(uint groupUin, uint sequence, uint random)
+    public async Task<bool> RemoveEssenceMessage(uint groupUin, ulong sequence, uint random)
     {
         var removeEssenceMessageEvent = RemoveEssenceMessageEvent.Create(groupUin, sequence, random);
         var results = await Collection.Business.SendEvent(removeEssenceMessageEvent);
