+++ Lagrange.Core/Internal/Packets/Service/Oidb/Request/OidbSvcTrpcTcp0xEAC.cs
@@ -9,7 +9,7 @@ internal class OidbSvcTrpcTcp0xEAC
 {
     [ProtoMember(1)] public uint GroupUin { get; set; }
     
-    [ProtoMember(2)] public uint Sequence { get; set; }
+    [ProtoMember(2)] public ulong Sequence { get; set; }
     
     [ProtoMember(3)] public uint Random { get; set; }
 }
